+++ flatbnb-angular/src/app/components/propietario/propietario-form/propietario-form.component.html
@@ -1,0 +1,1 @@
+<p>propietario-form works!</p>
