+++ flatbnb-angular/src/app/components/propietario/propietario-form/propietario-form.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-propietario-form',
+  standalone: false,
+  templateUrl: './propietario-form.component.html',
+  styleUrl: './propietario-form.component.css'
+})
+export class PropietarioFormComponent {
+
+}
