+++ flatbnb-angular/src/app/components/propietario/propietario-main-page/propietario.component.html
